#1181 - enhanced validation support for SlColorPicker / some cleanup
